fetch: updated how fetch works for log4cpp tarball packing; does not seem to affect other behavior.
Needs to be checked over more carefully with other recipes. Log4cpp
packs its stuff such that the 'tar tbzB 1' contains the directory name
as "./log4cpp" so the script thinks that the dirname is "./". Removing
this from the front means that start and end dirnames are the same
(log4cpp), so check to make sure before removing the original.